dev: rename ts_compiler to compiler
